reduct-rs: Add examples (#333)
* add example

* refactor creating bucket

* fix import

* clean code

* update readme

* update links

* simplify example

* run examples in pipeline

* update CHANGELOG

* don't publish if example broken

+++ reductstore/src/http_frontend/entry_api/read_batched.rs
@@ -136,7 +136,7 @@ fn fetch_and_response_batched_records(
                 if readers.is_empty() {
                     return Err(HttpError::from(err));
                 } else {
-                    if err.status() == ErrorCode::NoContent as i32 {
+                    if err.status() == ErrorCode::NoContent {
                         last = true;
                         break;
                     } else {
